updates to Delivery.kt

+++ app/src/main/java/com/example/inventorymanager/domain/model/Delivery.kt
@@ -24,5 +24,5 @@ data class Delivery(
     @PrimaryKey(autoGenerate = true)
     val deliveryId: Int,
     val saleDate: Date,
-    val customerId: Long  // Foreign key referencing Provider
+    val customerId: Int  // Foreign key referencing Provider
 )
